Minor build fixes - Add amalgamation to server-static - Only switch to gcc-4.8 if gcc is used - Build coveralls only if gcc is used - cast return value in examples (causes clang to fail otherwise)

+++ examples/client.c
@@ -253,6 +253,6 @@ int main(int argc, char *argv[]) {
 #endif
     UA_Client_disconnect(client);
     UA_Client_delete(client);
-    return UA_STATUSCODE_GOOD;
+    return (int) UA_STATUSCODE_GOOD;
 }
